Move tests for the DASD module
Move the unit tests for the DASD module to a new file.
